Adjust super calling to set non-primary creds
Because all Tempest scenario test cases have credentials set to
['primary'], the scenario tests' setup_clients() have cls.manager, going
on the assumption that primary creds are being used. Because of the use
of roles, cls.manager needs to be set to cls.os_roles_<rolename>. This
needs to dynamically happen on the base classes of the created classes.
There also had to be attribute setting done explicitly on the base class
in order to make it work.