Use exact role matching credentials provider
In the tempest preprovisioned credentials provider, when it looks for
users to get by role, it gets any user with those roles assigned. These
users may have extra roles added. Because we need to check against users
that have an exact specified role, we must be more restrictive in our
user selection.

More of the Tempest infrastructure needed to be swapped out in order to
return the new preprovisioned credential provider, and helper
methods/decorators were added appropriately to add that functionality.