Make everything work (for the most part)
This change brings in expected results, which is used to decide whether
or not a given test is expected to fail with a Forbidden error.

This change also fixed the credentials provider decorator to break the
infinite recursion that happened when chaining it with the role fan-out
decorator.

There were also just miscellaneous cleanups that were needed in order to
comply with tempest's calls.

What still doesn't work: In the fan-out decorator, it is not covering
the original class in the module, because doing so either causes
infinite recursion on super, or breaks the inheritance chain completely
(causing super to fail). Because of this, tests from the original test
class are still being run with the original credentials.