Use load_tests protocol to run dynamic tests
The load_tests protocol can be set on a module to allow custom loading
of tests. Because we do not want the original class's tests to be run
(because it is not being run with the proper roles), we can set the
tests to load to be only the tests in the classes we create.

The problem with this implementation is that now instead of skipping
tests when cls.skipException() is used, the tests now fail with the
SkipTest exception.