Make it all work (again)
So, I thought it was working last time I said it was working. It turns
out, somehow, it wasn't. So now I took an exact snapshot of having this
working with tempest.

I refactored things a little so now the test class being defined only
needs the one decorator, we automatically do the other decoration
within the bigger decorator.

The infrastructure for running py27 was also added, but none of the
tests are actually implemented yet.