Skip unlisted tests instead of passing
In order to prevent rug-pulling, we need to skip tests that are not
listed in the expected results file (or if a user is not listed on a
test). This way, if tempest adds a new test that we are not aware of, we
can just skip it instead of suddenly failing.