Dynamically load all tempest test cases
Now we can go searching through the entirety of tempest.api to find all
test classes. We then generate all of these classes to work with the
different role sets, and apply all of them to the
cinnamon_role.tests.api.tests_from_tempest module. The loading skips all
of the admin tests because that requires a different set of users.